APB-10575 organising packages
