+++ app/uk/gov/hmrc/agentregistration/shared/businessdetails/BusinessDetailsLimitedCompany.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2026 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.agentregistration.shared.businessdetails
+
+import uk.gov.hmrc.agentregistration.shared.BusinessType.LimitedCompany
+import uk.gov.hmrc.agentregistration.shared.BusinessType
+import uk.gov.hmrc.agentregistration.shared.CtUtr
+import uk.gov.hmrc.agentregistration.shared.SafeId
+import play.api.libs.json.Format
+import play.api.libs.json.Json
+
+final case class BusinessDetailsLimitedCompany(
+  safeId: SafeId,
+  businessType: BusinessType = LimitedCompany,
+  ctUtr: CtUtr,
+  companyProfile: CompanyProfile
+)
+
+object BusinessDetailsLimitedCompany:
+  given Format[BusinessDetailsLimitedCompany] = Json.format[BusinessDetailsLimitedCompany]
